# Remove the eviction timer after HTTP/1.1 downgrade