## Purpose

We have introduced the stale connection eviction for HTTP/2 server in [here](https://github.com/ballerina-platform/module-ballerina-http/pull/2066). So there is a eviction timer task runs for each `ServerChannelInitializer`. But this eviction timer task is not required if the connection got down-graded `HTTP/1.1`. This PR removes this tasks when the connection is down-graded.

> **Note:** we have observed an increase in the number of threads due to this eviction run tasks. In the case of HTTP/1.1 connections, this increase is very high which made the interceptor load test to have high error rate. Please refer to the results before and after this fix

Related to: https://github.com/ballerina-platform/ballerina-library/issues/6878

[Interceptor load-test results **before** this fix](https://github.com/ballerina-platform/module-ballerina-http/pull/2110/commits/49c5092abc54f3af5fb01d48848f38477b563e42?diff=unified&w=0):

| # Samples | Average | Median | Error % | Throughput | Received KB/sec | Payload | Users |
 | :-: | :-: | :-: | :-: | :-: | :-: | :-: | :-: |
 | 1408496 | 148 | 63 | **62.95%** | 404.7 | 88.6 | 50 | 60 |

[Interceptor load-test results **after** this fix](https://github.com/ballerina-platform/module-ballerina-http/pull/2112/commits/f847a351d2d891b970b4a9ff471d55f847601eea#):

| # Samples | Average | Median | Error % | Throughput | Received KB/sec | Payload | Users |
 | :-: | :-: | :-: | :-: | :-: | :-: | :-: | :-: |
| 216287 | 969 | 28 | **4.03%** | 61.7 | 8.9 | 50 | 60 |

**H2-H2 load test results with the `h2load` tool**

Number of maximum connections - 100
Number of maximum streams per connection - 10
Duration - 10 mins

Before the fix:
```console
finished in 600.05s, 18366.47 req/s, 630.69KB/s
requests: 11019885 total, 11020885 started, 11019885 done, 11019606 succeeded, 279 failed, 0 errored, 0 timeout
status codes: 11019606 2xx, 0 3xx, 0 4xx, 279 5xx
traffic: 369.55MB (387497892) total, 43.70MB (45826008) headers (space savings 94.93%), 136.67MB (143310954) data
                     min         max         mean         sd        +/- sd
time for request:      182us      31.03s     52.46ms    433.34ms    98.71%
time for connect:      972us      7.48ms      2.96ms      1.34ms    78.00%
time to 1st byte:   591.38ms       2.39s    879.74ms    322.20ms    85.00%
req/s           :     133.19      236.24      183.65       18.79    72.00%
```

<img width="1403" alt="Screenshot 2024-08-16 at 08 56 34" src="https://github.com/user-attachments/assets/80913a4f-d295-4dbe-bc59-aa76b53dbe19">

After the fix:
```console
finished in 600.04s, 20822.41 req/s, 714.75KB/s
requests: 12493446 total, 12494446 started, 12493446 done, 12493232 succeeded, 214 failed, 0 errored, 0 timeout
status codes: 12493232 2xx, 0 3xx, 0 4xx, 214 5xx
traffic: 418.80MB (439142797) total, 49.40MB (51802739) headers (space savings 94.94%), 154.93MB (162455030) data
                     min         max         mean         sd        +/- sd
time for request:      169us      30.39s     47.74ms    297.51ms    98.36%
time for connect:      292us      2.95ms      1.72ms       747us    66.00%
time to 1st byte:   478.88ms       2.95s    847.14ms    742.62ms    86.00%
req/s           :     166.05      267.39      208.21       20.79    68.00%
```

<img width="1403" alt="Screenshot 2024-08-16 at 08 56 54" src="https://github.com/user-attachments/assets/b82bb3f3-5d78-4f16-8b21-1abcc3224727">

With Ballerina SwanLake 2201.9.2:
```console
finished in 600.05s, 18295.76 req/s, 628.48KB/s
requests: 10977457 total, 10978457 started, 10977457 done, 10976282 succeeded, 1175 failed, 0 errored, 0 timeout
status codes: 10976282 2xx, 0 3xx, 0 4xx, 1175 5xx
traffic: 368.25MB (386136304) total, 43.50MB (45611237) headers (space savings 94.93%), 136.31MB (142927841) data
                     min         max         mean         sd        +/- sd
time for request:      181us      30.59s     50.99ms    478.69ms    98.98%
time for connect:      484us      8.07ms      2.74ms      3.14ms    74.00%
time to 1st byte:   544.06ms       2.49s    961.07ms    541.89ms    90.00%
req/s           :     146.16      240.49      182.95       20.17    70.00%
```

<img width="1403" alt="Screenshot 2024-08-16 at 08 57 19" src="https://github.com/user-attachments/assets/691d522f-35f6-444b-a713-d6b3f5e53876">

## Examples

N/A

## Checklist
- [x] Linked to an issue
- [ ] ~Updated the changelog~
- [ ] ~Added tests~
- [ ] ~Updated the spec~
- [ ] ~Checked native-image compatibility~
- [ ] ~Checked the impact on OpenAPI generation~
- [x] [Check downstream modules](https://github.com/ballerina-platform/module-ballerina-http/actions/runs/10415256115)